cache: full cache update functionality
We of course need a way to bypass the cache,
with a `nocache` parameter to `get`.

We need to tighten the tzawareness of rev/insertion dates.

For the datetime lisp expression, we provide a built-in `now` variable
which produces the current tzaware timestamp (in utc).

For the next steps, we need to address:
* history
* insertion_dates