cache/refresh: do no bypass the caches of the components of a formula
We want to:
* bypass the cache of the current formula
* but not the cache of underlying formulas

By using directly `eval_formula` we exploit the `stopnames` internal
expansion logic which treat formula with a cache as a primary series
(which get + nocache completely bypassed).

In tshistory_refinery/tsio, we indeed use the stopnames parameter to
stop the formula expansion on series that have a cache.