cache: simplification
The notion of a status for the policy itself, signalling its
readiness, is wrong. We remove it.

What remains is a "ready" flag on series signalling its cache being
upgraded or not.

We maybe want an "initial" flag, but per series.

The refresh_now function receives a fix in the form of looking at the
latest known value date to compute a better from_value_date.

We also remove the api/refresh_now test, for this thing is not really
testable.