cache/series_lock_ready: replace the implementation with something more robust
The purpose of this is to avoid competing series caches refreshes.

The current implementation is vulnerable to hard crashes (tasks being
killed is far from being unheard of).

By using a postgres transactional advisory lock, we delegate the issue
to the transaction manager. In case of a crash, the lock should be
always released.

To identify the unicity of the lock, we compute a number from the
series name. The chosen method does not garantees absolute unicity but
avoid building too big numbers.

closes #3